login + authentication

+++ routes/web.php
@@ -21,28 +21,34 @@ use App\Http\Controllers\ListingController;
 Route::get('/', [ListingController::class, 'index']);
 
  //Show Create Form
- Route::get('/listings/create', [ListingController::class, 'create']);
+ Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
  //Store Listing Data
- Route::post('/listings', [ListingController::class, 'store']);
+ Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
  //Show Edit Form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 //Update Listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 //Delete Listing (destroy or delete or Petras - no matter but the same must be on controller)
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
  //Single listing ammended with 404 and it must be below a create and store
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 //Show Register/Create Form php artisan make:controller UserController
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 //Create New User
 Route::post('/users', [UserController::class, 'store']);
 
 //Logout User
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+//Show Login Form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+//Log In User
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
